requirements.txt
crewai
autogen
streamlit

numpy
Pillow
Requests
PyPDF2